move decoder onto model class
